+++ src/main/kotlin/de/moritzf/opencodewebpanel/toolWindow/OpenCodeWebToolWindowFactoryImpl.kt
@@ -46,6 +46,7 @@ class OpenCodeWebToolWindowFactoryImpl : ToolWindowFactory, DumbAware {
                 OpenCodeZoomOutAction(),
                 OpenCodeZoomInAction(),
                 OpenCodeResetZoomAction(),
+                OpenCodeAutoAcceptPermissionsAction(),
                 OpenCodeReloadPageAction(),
                 OpenCodeRestartServerAction(),
                 OpenCodeResetWebStateAction(),